+++ ios/Classes/EidmsdkPlugin.swift
@@ -70,7 +70,7 @@ public class EidmsdkPlugin: NSObject, FlutterPlugin {
   }
 
   public func showTutorial(result: @escaping FlutterResult) {
-    eidHandler.showTutorial(from: findViewController(), environment: .minvProd,) {
+    eidHandler.showTutorial(from: findViewController(), environment: .minvProd) {
       result(nil)
     }
   }